Factor out common window_getchar.
